Added Title Banner and Nav Bar Added Title Banner using a temporary image that will be replaced with a render of the spacecraft in the future. I made the text for the title overlay the image. I also allowed the image to be dimmed using a '::before' pseudo-element to overlay a mostly transparent black image over the banner image but not the text.
Added a Nav Bar with stub links to the other pages that will be added in the future. I used a flexbox to align the 'about' and 'contact' links to the right of the page. I aligned the 'Mesh Merchant' text (name of the page) to the left hand side and added a link to direct the user back to the main page. I also added a hover effect to the links to make them lighten in colour and bold when the mouse hovers over them. To prevent the text from shifting when hovered due to their relative sizes changing when bolded, I faked the bold effect by adding a text shadow that is the same colour as the text. This solution is a bit hacky but it works well.